Move points and vectors tests to separate file
