+++ tests/expectedoutputs/TestPointsAndVectors/test_WMSGetMap_Arrows_example_windbarbs_on_gridded_netcdf.png
