+++ tests/expectedoutputs/TestPointsAndVectors/test_WMSGetMap_Barbs_example_windbarbs_from_pointdata_csv_different_style_options.png
