+++ tests/expectedoutputs/TestPointsAndVectors/test_WMSGetMap_Barbs_example_windbarbs_on_gridded_netcdf.png
