+++ tests/expectedoutputs/TestPointsAndVectors/test_WMSGetMap_Discs_example_windbarbs_from_pointdata_csv.png
