+++ tests/expectedoutputs/TestPointsAndVectors/test_WMSGetMap_Discs_example_windbarbs_on_gridded_netcdf.png
