+++ tests/expectedoutputs/TestPointsAndVectors/test_WMSGetMap_KNMI_WebSite_AnimatedGifImagery_temperature_style_temperature_thinned.png
