+++ tests/expectedoutputs/TestPointsAndVectors/test_WMSGetMap_select_temperature_as_point_from_grid.png
