+++ tests/expectedoutputs/TestPointsAndVectors/test_WMSGetMap_wave_direction_vector.png
