+++ tests/expectedoutputs/TestPointsAndVectors/test_WMSGetMap_wave_direction_vector_with_add_dataobject.png
