+++ tests/expectedoutputs/TestPointsAndVectors/test_WMSGetMap_windbarbs_kts_selectpoint_for_grids.png
