+++ tests/expectedoutputs/TestPointsAndVectors/test_WMSGetMap_windbarbs_on_gridded_netcdf_striding_offset.png
